(refactor) Home components and created League Info page/route
