+++ src/views/LeagueInfo.jsx
@@ -4,6 +4,8 @@ function LeagueInfo() {
   return (
     <div>
       <h1>League Info</h1>
+      <p>Starting Bank Roll</p>
+      <p>Length of Match (day, week)</p>
     </div>
   );
 }
